requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
pytz==2019.3